requirements1.txt
aitertools==0.1.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.2
distro==1.9.0
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
greenlet==3.0.3
groq==0.4.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.4
openai==1.12.0
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sse-starlette==2.0.0
starlette==0.35.1
tqdm==4.66.1
typing_extensions==4.9.0
uvicorn==0.27.0
wheel==0.41.2
